requirements.txt
Django==3.1.6
django-model-utils==4.1.1
djangorestframework==3.12.2
djangorestframework-xml==2.0.0
django-thumbnails==0.4.0
django-compressor==2.4
django_extensions==3.1.0
djangorestframework_simplejwt==4.4.0

psycopg2-binary==2.8.6
time-machine==2.0.1
lxml==4.6.2

# phone number
id-phonenumbers==0.4.0
phonenumberslite==8.12.17

responses==0.12.1
requests==2.25.1

redis==3.5.3
pillow==8.1.0
django-redis==4.12.1
django-libsass==0.8

# raven==6.5.0
furl==2.1.0

# timehelper
python-dateutil==2.8.1

# sentry
sentry-sdk==0.19.5

PyJWT==1.7.1
